doesn't run just update

+++ src/JsonHandler.h
@@ -1,11 +1,12 @@
 #ifndef JSON_HANDLER_H
 #define JSON_HANDLER_H
 
-#include <ArduinoJson.h>  // Include the ArduinoJson library
+#include <ArduinoJson.h> // Include the ArduinoJson library
 
-class JsonHandler {
+class JsonHandler
+{
 public:
-    static bool parseJson(String jsonString, int &gps_lat, int &gps_lng, int &wind_direction_raw);
+    static bool parseJson(String jsonString, float &gps_lat, float &gps_lng, int &wind_direction_raw);
 };
 
 #endif // JSON_HANDLER_H
